Use Custom L2 token and bridge
This modifies SESH to have a custom bridge so we can have an upgradable
L2 token contract. This provides the necessary code so that the arbitrum
bridge can mint/burn tokens on the L2 token as necessary, and makes it
possible to register the tokens with the arbitrum router and gateway.

To use these we call the following scripts in this order
- deploy-l1-token
- deploy-l2-token
- register-and-bridge

with the token deployment addresses needing to be added to the scripts
as they are deployed.