addBLSPublicKey: check required operator stake
The operator stake requirement is currently checked only in the
MultiContributor, but not in addBLSPublicKey itself, and that means that
you could add a key to the contract with a too-small operator staking
requirement, which oxend will ultimately refuse as invalid.  This fixes
it by checking in addBLSPublicKey.

Also removes the `FirstContributorMismatch` event that hasn't been used
in a while.